+++ Tests/LinuxMain.swift
@@ -1,7 +1,13 @@
 import XCTest
-import XCTestManifests
 import SwiftCueSheetTests
 
-var tests = allTests()
-
-XCTMain(tests)
+public func allTests() -> [XCTestCaseEntry] {
+    return [
+        testCase(ParsingFailTests.allTests),
+        testCase(TimeTests.allTests),
+        testCase(TrackEditTests.allTests),
+        testCase(CueSheetTests.allTests)
+    ]
+}
+//var tests = allTests()
+XCTMain(allTests())
